+++ src/app/common/base/base-entity.model.ts
@@ -2,5 +2,4 @@ export interface IBaseEntity {
   id?: string;
   readonly createdAt?: Date;
   readonly updatedAt?: Date;
-  readonly deletedDate?: Date;
 }