+++ src/app/common/interface/filter-field.interface.ts
@@ -1,8 +1,0 @@
-import { QueryNarrowingOperators } from '@common/enum/query-operators.enum';
-
-export interface IFilterField<T> {
-  name: string;
-  target?: keyof T;
-  relationTarget?: string;
-  operation: QueryNarrowingOperators;
-}